Рефакторинг файла context.js #54
- оставил в нём только то, что касается непосредственно контекста
- остальное перенесено в константы - const.js
